Move game roster actions into slice logic file
